Support Generation of Web Connector's QWC Files
Add a QBWC module with a QWCConfig type & a generateConnectorFile
function to turn the type into the XML for the Web Connector's QWC
Configuration files. Associated types are also defined in this
module(e.g., AuthFlag, QBType).

The QWC file is what a web service provides to it's user so they can
pair their QuickBooks Desktop's company file with the web service.