Add CertURL QWC Option & Route for Verifying Certificate
Add an optional CertURL field to the generated QWC file. This informs
the WebConnector of the URL to test/verify the SSL certificate of the
server. This is needed since the syncing route only accepts POST
requests, while the WebConnector uses a GET request for ceritifcate
verification.

Add a `/cert/` Route that accepts GET requests and returns nothing
solely for the QBWC to verify the server's HTTPS certificate.