Support the ClientVersion Callback
Add the ClientVersion & ClientVersionResp values along with
FromXML/ToXML instances for responding to the clientVersion QBWC
callback.

The syncing route simply ignores the version of the client and always
accepts the connection. In the future we may parse the version string
and only support versions within the range that the server has been
tested with.