Add Application-Wide Configuration Datatype
Add an AppConfig type to the server for static application configuration
options such as TLS settings, the server host/port, & the QBWC hostname
and sync interval/UUID.

Currently the QBWC configuration options are not used - this will be
integrated when we write a custom Monad stack for the servant routes.
But the TLS & web server settings are used in the server's Main module.

Add options to disable TLS, allow insecure requests when using TLS, and
specify the certificate/key file names.