Provide App Configuraton to Server Routes
Add an AppM type for our server route handlers. This is a ReaderT
storing the AppConfig over the IO monad. It derives MonadReader &
MonadIO typeclasses & will store a database connection pool when we
integrate the server with a database.

Add `username` & `password` fields to the `account-sync` field of the
settings file. The username defaults to `account-sync` but the password
must be supplied via a user's `settings.yaml` file.

Access to the AppConfig lets us generate the QWC file using the hostname
& username from the settings files, as well as verify the password
during authentication.