Integrate Server with a PostgreSQL Database
Add a new AppEnv type for the application that contains the static
AppConfig configuration data and a Pool of database connections -
modifying the AppM monad to have access to this new type instead of just
the AppConfig.

Add an AppSqlM type to represent the database querying monad, along with
a runDB function to run an AppSqlM value in the AppM monad. Add a
MonadUnliftIO instance to the AppM type to allow database operations in
the AppM monad; the instance is a simple wrapper around the instance for
AppM's inner ReaderT.

Add new fields to the AppConfig type representing the database
connection details and add a buildConnectionString function to the
Config module that uses these details to build the postgresql connection
string(where blank values use postgresql's defaults).

Add DB.Schema & DB.Fields modules containing the database table schema
as well as custom field types. Currently we have a single table that
tracks WebConnector sessions along with their UUID ticket, session type,
current status, and potential errors the session encountered. This
includes new types for the SessionType, SessionError, & SessionStatus.

Modify the Main module to create a database connection pool & an AppEnv
value, passing these to the server.

Modify the Authenticate callback handler so that sessions are saved to
the database, along with an InvalidAuthentication SessionError if
authentication failed. Add a generateUniqueTicket function for ensuring
that newly generated ticket UUIDs do not already exist in the database.