Add Parser Module for Simplifying XML Parsing
Add a Parser module with a Parser type, runner function, and
selection/reading functions.

Add a Response type to the QBXML module along with a HostData type,
representing the data returned from a qbXML HostQuery.

The Parser type is a MonadReader with a cursor on the current Element.
The selection functions can descend into the Element's children & the
reading functions can parse the text content of the current Element.
These can be combined in a monadic fashion, preventing us from having
the repetitive if/case/parse style used in the WebConnector module's
parsing functions.

Currently the Parser is only used for parsing the HostData, although it
will probably be re-used for the WebConnector module's parsers and the
FromXML typeclass.

Only simple error handling(of Strings) is supported. Eventually the
parser will keep track of the current Element depth(e.g., "html > body >
div") for better error messages.