Add Custom Context & Error Type to Parser Module
Add the ParserContext & ParsingError types. The context keeps track of
the element names traversed up to & including the current Element. This
context is used when using ParsingErrors, allowing us to keep track of
the path of element names we've followed when reaching an error.
ParsingErrors are rendered using this hierarchy along with the specific
error message.

Eventually, we'd like to try multiple parsers & either return the first
success, or find the longest error and return that instead, as well as
provide concrete datatypes for every expected error type instead of Text
values.