Remove MonadError & MonadReader Instances for the Parser Type
Remove the MonadError ParsingError & MonadReader ParserContext derived
instances for the Parser type.

Instead of allowing arbitrary typeclass function calls, we expose our
own context querying, descent, & error throwing functions. The descent
function maintains the `hierarchy` field in the ParserContext as we
descend into Elements & the parseError function uses this hierarchy to
build the ParsingError that is thrown.

Essentially, parseError is a specialized `throwError` function & descend
is a specialized `local` function.