Integrate Parser Module with FromXML Typeclass
Modify the FromXML typeclass so that the fromXML method takes no
arguments and simply returns a `Parser` for the instance type.

Modify the SOAP & XML modules to use these parsers to decode request
bodies & modify the WebConnector module so that the Callback type's
FromXML instance returns a Parser.

Add the `withNamespace` function to the Parser module to ease creation
of Text -> Name functions with predefined namespaces, such as the
WebConnector.qbName & SOAP.soapName functions.

While this commit does not move the parsers from the QBXML module into
FromXML instances, this is our eventual plan.