Refactor New Company Form, Log State to Console
Modify the form to log the NewCompany.State to the console on
submission.

Wrap the New Company inputs in a form element, changing the type of the
button to "submit" & making all inputs required.

Add PreventDefaultSubmit & LogToConsole typeclasses to the App module
for preventing the default submission event & logging data to the
console.