Rework Typeclasses to Remove MonadEffect Constraints
The way we were defining typeclasses was requiring a MonadEffect
constraint on functions that used their methods. This is because
instance matching does not limit by constraints on the instance and
instead requires those constraints. So `MonadEffect m => Class m` says
that all `m` that utilize `Class`'s methods should provide a
`MonadEffect` as well.

To overcome this, we change the `MonadEffect m =>` instances to match
our `AppM` directly, and write a wrapper for the `H.HalogenM` type that
the component & parentComponent functions return.

This allows us to remove all the MonadEffect constraint from all of our
component functions.