Refactor the QWC File Route, Add JSON Encoding
Modify the `/qwc/` route to return a new QWCFile type. This type
supports both XML & JSON requests, so the frontend can fetch the
contents as JSON & present a download link to users.

The main purpose is to introduce the QWCFile type that we can return
when creating a new company. The qwc route needs to be modified to take
a company ID so it can correctly generate the appropriate config file,
but this isn't necessary until the frontend adds a Companies or Status
page.