Add Module for Validating API Types
Add the Validation module which contains error types, validators, and
validation runners. Validators accumulate errors by field, and running a
validation will cause servant to throw a 422 error with the errors
encoded as JSON in the body contents.

The utility functions and tuples can help construct generic validators.
For example, we do this when validating name & username uniqueness in
the newCompany route.