Return QWC File From New Company Route
Instead of return a unit value when creating a new Company, return the
generated QWC File. There are some global values we will add to the
final qwc file, such as a configurable application name & base URL for
the API.

The qwc route still does not return a company-specific file.

Modify the validationError function so it always throws an error. This
allows the whenValid function's onSuccess argument to return arbitrary
values instead of the unit value.