Show QWC Download Button After Creating Company
When a company creation request is successful, change the NewCompany
page to some descriptive text on how to initialize the web connector as
well as a download button for the new company's qwc file.

Add a ManageObjectURLs typeclass for creating/revoking Object URLs from
Blobs.

Add styles for successful form submission text.

Add a QWCFile type & decoder that wraps a Blob.

Modify the newCompanyRequest Server method to return the company's QWCFile.