Rework Authentication Callback & QWC Route
Rework the Authentication callback handler so that it uses the company
credentials from the database instead of static values in the
configuration file. This lets us sync the accounts of multiple company
files. Valid authentications will save the company id in the ticket for
use in future callacks. Sessions that have just completed authentication
are marked with the Authenticated status. If we know the path to the
Company File, we try to connect directly to it. Otherwise we're probably
doing the initial sync for a new company & need to have the file open
first.

Rework the `/qwc/` route so it takes a Category ID(`/qwc/<category-id>`)
and generates the correct QWC file. We will do more customization in the
future, like a base URL for api requests(e.g., `/api/`) and custom
application names.

Remove the Username & Password fields from the configuration file &
type.

Add an Authenticated value to the SessionStatus type.