Validate Session & Store Filename in sendRequestXML Handler
During the InitialSendRequestXML callback, ensure the session for the
passed ticket has a Company ID. This ensures that authorization was
sucessful for the session.

When the Company does not have a stored Company File, save the path
passed by the Web Connector.

Before sending the AccountQuery to the WebConnector, update the
session's status to 'RequestedAccounts'.

Add an Either () branch to the SendRequestXMLResp value. The Left value
indicates an error and that no qbXML should be sent to the WebConnector.