Add the closeConnection Callback & Handler
Add a CloseConnection Callback value & a CloseConnectionResp
CallbackResponse value. The callback's only parameter is the session
ticket. The response contains a string to display in the UI of the
WebConnector.

The handler returns a success message if the ticket's session has no
error, otherwise it shows the error. The other possible message is if
the given ticket does not exist in the database, making the session
unidentifiable.