Add Application Name & Base URL to Application Config
Add a `name` key to the config file for the name to present to users of
the application.

Add a `base-url` key for prepending to API server URLs in the QWC files.
This is useful if you are serving the frontend and proxying the backend
behind some path, like `/api/...`.

Only include the port number in URLs if it is not port number 80.