Sync QuickBooks Accounts to Database
Add an Account datatype & table which we use to store the details of a
Company's Quickbooks Account.

Modify the ReceiveResponseXML callback handler to update our versions of
the accounts from the AccountQueryResponse from the WebConnector. We
fold over the returned accounts, creating a list of root
accounts, a map from references to the returned account data, and a map
from references to child references. We then insert/update accounts by
following every level of the parent->children map for each root account.
Upon sucessful update, the Company's last sync time is set to the
current time. However, the sync time is not currently used to filter the
returned Accounts.

Add an `UpdatingAccounts` value to the `SessionStatus` type.

Add a PersitField instance for the AccountType field. This is put in the
DB.Fields module as an orphan instance since we may pull out the
quickbooks types into a seperate package. A better option would be a
newtype wrapper around the AccountType field instead of an orphan
instance.

Add Generic & Hashable instances for the ListReference type, allowing us
to use them as keys in a HashMap.