Add Optional ModifiedTime Filter to AccountQuery
Change the AccountQuery QBXML.Request value to take an optional
AccountQueryFilters value, which contains a UTCTime that indicates we
only want the AccountData for accounts after that time.

Use this modification time filter in the sendRequestXML callback when a
company has a LastSyncTime specified, so we only pull data that has
changed since the last sync.

Unfortunately the last modified time in QuickBooks ignores the DST
setting of the operating system so we must adjust the LastSyncTime
during daylight savings time before we send it to QuickBooks. For
example, if we modify an account at 8:30pm UTC during DST, QuickBooks
will record that the account was modified at 9:30pm UTC & will return
that account in every sync until the actual time hits 9:31pm UTC.

It should be noted that QuickBooks does not update the modified time for
an account when it's FullBalance changes, thus parent accounts will not
be returned during ModifiedTime queries if the Balance for an account
has changed. We do not currently store the FullBalance or Balance
fields, but this may require some reworking when we want to show
balances on the chart of accounts.