Add ModifiedTime Field to Accounts & Remove FullName Field
Add the accountModifiedTime field to the Account type, grabbing the
value from the AccountData in the WebConnector Sync route.

Remove the FullName field. This is simply a colon seperated list of the
names of the Account's predecessors. Unfortunately, modifying a parent
account's name does not return the children with updated FullNames from
an AccountQuery with the ModifiedTime filter, so the FullNames will go
out of sync with the actual values in QuickBooks.