Fix Partial Account Updates
Fix a bug where non-root accounts would not update after the initial
update. This occurred because we upsert accounts from the root node down
to each level of child nodes so that we always have every child's parent
ID. However, the ModifiedTime AccountQuery filter does not return the
parent's of every modified account causing those accounts to not be
updated.

To fix this, we first try upserting using the root->child tree,
returning the list of references that were updated. We then determine
which references were not updated and use a seeking/querying algorithm
to upsert those. The algorithm first searches for the parent's
in the list of remaining references to process - if found it upserts the
parents first followed by the child. If not found, it queries the
database for the parent and uses it's ID to upsert the child. Then
processing continues on the remaining accounts.