Add Type to Describe Different XML Parsing Errors
Add the ParsingErrorType with tags for the different possible XML
parsing errors we run into, as well as a GenericParsingError that takes
a simple message, retaining the previous behavior.

Add a throwParsingError function for throwing ParsingErrors from a
ParsingErrorType. Modify the parseError function to throw a ParsingError
using the GenericParsingError type.

Restructure the ParsingError type, removing the message field and adding
errorType and errorCursor fields. Derive the Show instance instead of
manually defining it.

Add a prettyParsingError function for pretty-printing ParsingErrors.
This has the same overall structure as the old Show method for
ParsingErrors, but includes more information about the error when the
errorType is something other than a GenericParsingError.