Use Typeclasses to Limit Effects in Route Handlers
Add various typeclasses to the Types module to describe the effects a
function can have. This allows us to remove the `AppM` from a function's
return type & use a generic `m` that is constrained by the typeclasses
which prevents arbitrary IO effects. Examples of effects are executing
DB actions & fetching the current time.

Refactor the handlers to the use the new typeclasses.