Refactor the Label/Help/Input Element Alignment
Instead of adding some padding to the top of the label/help element,
simply push the input to the bottom of the row.

Don't include the paragraph tag when the help text for an input is
blank. This ensures the label is properly centered vertically, w.r.t.
the input element.