Add UserName Field to QWCConfig Type
Originally a QWCConfig was separate from the User the config was meant
for. This commit refactors the QWCConfig type, XML generation, &
relevant routes to integrate the username directly into the QWCConfig
type.

This means we no longer need the `QWCFile = (QWCConfig, Text)` newtype
wrapper to represent a complete QWCConfig, we can simply return a
QWCConfig from our route handlers.