Touch Up Styling of Form Inputs
Add an inset box shadow to non-error & non-focused input elements. This
gives them a little bit of depth instead of looking completely flat.

Add styles for disabled inputs. They now have a grayed-out background &
lighter font color, and show a "Not Allowed" cursor when hovered over.

Ensure that all form inputs use the same font as the rest of the page.
This was needed since Date inputs default to a monospaced font.