Add Entry Fields to the New Trip Form
Add the basic entry-specific fields for the new trip form. This includes
the date, tripper name, trip number, advance, & cash returned. There is
also a disabled Cash Spent field which is automatically calculated from
the trip advance & cash returned.

The date & trip number are automatically filled out using the current
day. The trip number is generated using the trip date. When the date is
changed, the trip number will automatically be updated if it's current
value was automatically generated.

This does not include any Store tables with Transaction fields, or the
Company field which will determine the accounts presented in the form.
There is currently no validation, and the form is simply printed to the
console when submit instead of being POSTed to the server.

Add a DateTime typeclass to the App module. This is used for pulling the
current DateTime and parsing Date values from Strings.