Fetch Companies When Visiting Add a Trip Page
Fetch all available Companies(IDs & names) when the NewTrip route is
initialized. Store these in the page's State.

We should probably wrap this in a Maybe, showing a loading message while
the value is Nothing, an error message when it's an empty list, and the
form when there are available companies.