Add a Company Dropdown to the Add a Trip Page
Add a select element for picking the Company for a Trip to the NewTrip
page. Selecting a company will fetch the Accounts for that Company. When
initializing the page, the first Company returned will be selected & the
Accounts will automatically be loaded. The selected Company is stored in
the page's State.

Add an accountsRequest method to the Server typeclass for fetching the
AccountData for a Company by it's ID. Add an AccountData type for the
result of the request, containing an Account's ID, name, & description.

This will enable us to build an Accounts dropdown for the Transaction
rows in the TripStop tables.