Add Trip Stops & Transaction Tables to New Trip Form
Add types for TripStops & Transactions to the NewTrip page.

Trip stops are rendered as fieldsets and have names, receipt totals, and
Transactions. Transactions are rendered as tables and contain an
Account, Purchase Description, Cost, Tax %, Total Cost, & whether they
were a purchase or a return.

We currently hardcode the default tax rate to 5.3% to match VA's sales
tax, but eventually this will be a configurable dropdown/input. When a
number is entered into the Cost field for a transction, the total cost
will be automatically calcualted using the cost & tax percentage. The
total costs for each transaction in a Trip stop is summed & displayed
below the table body, along with the Out of Balance amount between the
transaction sum & the "Total Spent" field of the Trip stop.

Transaction rows have a link in a cell for deleting their row from the
TripStop, and there is a link below the table body for adding more
Transaction rows.

There is a button for removing a Trip Stop from the form, as well as a
button for adding new Trip Stops to a form.

Currently the Transaction table is unstyled & there is no Account
dropdown. We will use a separate library for the accounts dropdown, and
then style the table inputs similarly to the normal form fields.

A Total & Out of Balance counter will be added below all TripStops that
sum their Total Spent field & calculate the difference from the Cash
Spent field.

Right now, you must tab through all the table inputs, but we will add
the ability to press Enter to move to the first input in the next row of
the table. On the last row, this should add an additional row of inputs.

No validation is currently done on the Trip Stop. We may do client-side
validation, or just pass it through to the server for validation.