Add Typeclasses for Preventing Enter Actions & Focusing Elements
Add the PreventDefaultEnter typeclass, which takes a KeyboardEvent and
prevents the default action if it's a keydown event for the Enter key,
returning whether or not the default action was prevented.

Add the FocusElement typeclass, which takes an HTML Element & focuses
it.

These are implemented for the AppM & HalogenM types.