Move Down a Row in the Transaction Table When Pressing Enter
Add keydown event handlers to each input in the Transaction tables. When
the Enter key is pressed in one of these inputs, the focus will switch
to the first input of the next Transaction row. If the focus is already
in the last row, a new row will be added to the table & it's first input
will be focused.