Add Autocompleting Select Component for Accounts
Add an AccountSelect module containing a component that takes an array
of accounts and renders an input field with a dropdown menu that lets
users search for & select accounts. This support keyboard navigation,
mouse selection, basic filtering, & selection clearing. It also provides
a Message & Query to facilitate focusing the next row's AccountSelect
when we render them in table rows.

Add styles for the new component.

Add a SelectComponent typeclass to the App module for components that
render child NSelect components.