Add AccountSelect Components to the New Trip Page
Add the new AccountSelect component to the Transaction tables of the New
Trip page.

Change the account field in the Transaction type to store a potential
AccountData instead of a raw String.

Adds a `ChildSlots` type for designating the account select components.
Note that this will need refactoring - using array indexes for the slot
values means that removing rows always drops the last row! Instead we
need to implement counters for TripStops & Transactions so newly added
Transaction table rows & their AccountSelect components have unique
indexes.

Modify the "Enter moves down a row" functionality in the table inputs so
that the AccountSelect component is focused instead of the details
field.