Fix Visibility of Highlighted Account When Opening Select Dropdown
Fix a bug causing the highlighted item to be hidden every other time the
AccountSelect dropdown was opened. This occured because we change the
height of the dropdown for a nice transition effect instead of simply
adding/removing the dropdown element.

The fix involves manually re-setting the highlighted element when the
dropdown becomes visible. This triggers the Select component's
functionality for ensuring the highlighted item is within view.