Highlight Matches in the AccountSelect Dropdown
Use the fuzzy package to do fuzzy searches of Account names &
descriptions, instead of simply doing case-sensitive substring matching.

The filteredItems array is now built from a new SelectItem type which
bundles an AccountData with a matching result(for non-empty search
values).

Dropdown items are kept if they have 11 out of every 12 characters from
the search string & in descending order of match strength. This may need
to be tweaked once we have our real set of Accounts to deal with.

Search matches are highlight in the dropdown items by encasing the
matches in spans & styling them to a dark text color & yellow
background.