Show Account Types in Account Select Dropdowns
Modify the AccountSelect component, rendering the AccountType in the
dropdown items & include the account type in the searched values. The
rendered Account Type appears before the description text, and the two
are separated by a colon when description text exists. The styling is
similar to the description text, except it is not rendered in italics.

Add an AccountType datatype representing all of the QuickBooks Account
types, along with a prettyAccountType function to stringify them into
base categories of Assets, Liabilities, Equity, Cost of Goods Sold,
Income, & Expenses.

Modify the AccountData type to include field for the AccountType and
parse it from the API's JSON response.