Don't Attempt Matching Against the AccountType in AccountSelects
Remove the AccountType matching for the dropdown items in AccountSelect
components.

This was messing up the ordering because the fuzzy package doesn't
support matching with arbitrary word orders, so searches like "auto
expense" were causing large distances in the expected matches.

Reduce the match threshold to accept any match of at least 1 character.