Properly Handle Removal of Stops & Transactions
Change the slot indentifier of the AccountSelect components to use a
Tuple of a new StopCount & TransactionCount types instead of their Array
indexes. These types are wrappers around an Int, but stay static for
each Stop & Transaction instead of changing when their underlying array
is changed. This persists the state of the AccountSelect components with
respect to each Stop's Transaction, instead of the previous behavior
where removing a row would make the new Transaction positions inherit
the AccountSelect components of the previous Transactions in that
position.