Use Custom Checkbox Inputs With Thematic Styling
Wrap the isReturn checkboxes on the New Trip page in a label, with an
additional custom-checkbox span.

Add styles for the custom checkbox elements. We make the checkbox inputs
transparent and style the span inside the label instead. This keeps the
event handlers working properly & allows us to customize the styling of
the checkboxes, which isn't possible with browser's native checkbox
elements.