Use More Intelligent Out-of-Balance Coloring
Modify the TripStop's Out-of-Balance counter highlighting so that there
is only a highlight applied when the stopTotal and/or any transaction
has an amount entered.

Factor out functions for calculating a stop's Transaction total, the
trip's cash spent amount, and the Out-of-Balance highlight color class.

Allow the out-of-balance coloring to apply to any elements with the
given classes instead of restricting them to rows in a table footer.