Handle Submission of New Trip Form
Instead of simply logging the state to the console when the NewTrip form
is submitted, validate the state and submit the data to the server if
the form is valid.

Currently nothing is done with the returned TripId. Once the TripDetails
page is implemented, a successful submission will redirect to the details
page of the newly added Trip.